requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
bcrypt==4.1.2
email-validator==2.1.0
Werkzeug==3.0.1
gunicorn==21.2.0
python-dateutil==2.8.2
